requirements.txt
pandas>=2.0,<3
sqlalchemy>=2.0
python-dotenv>=1.0
tqdm>=4.0
python-dateutil>=2.8
streamlit>=1.28
plotly>=5.18
scikit-learn>=1.3
matplotlib>=3.8
seaborn>=0.13